requirements.txt

Django==3.1.5
django-crispy-forms==1.10.0
mysqlclient==2.0.3
Pillow==8.1.0
